added page and layout outside locale folder
