+++ src/app/layout.tsx
@@ -1,38 +1,12 @@
-import type { Metadata } from "next";
-import { inter } from "./fonts";
-import "./globals.css";
-import Navbar from "@/components/common/navbar";
-import Footer from "@/components/common/footer";
-import { ThemeProvider } from "next-themes";
+import { ReactNode } from "react";
 
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+type Props = {
+  children: ReactNode;
 };
 
-const RootLayout = ({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) => {
-  return (
-    <html lang="en" suppressHydrationWarning>
-      <body className={inter.className}>
-        <ThemeProvider
-          attribute="class"
-          defaultTheme="system"
-          enableSystem
-          disableTransitionOnChange
-        >
-          <div className="flex min-h-screen flex-col">
-            <Navbar />
-            <div className="flex-1">{children}</div>
-            <Footer />
-          </div>
-        </ThemeProvider>
-      </body>
-    </html>
-  );
+// Since we have a `not-found.tsx` page on the root, a layout file
+// is required, even if it's just passing children through.
+const RootLayout = ({ children }: Props) => {
+  return children;
 };
-
 export default RootLayout;
